// server-v2/src/types/percent/user/updateUser/body.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface UpdateUserBody {
  name?: string;
  profile?: string;
  [k: string]: unknown;
}